Migration added, connection with the database established.

+++ NLayer.Core/Product.cs
@@ -11,7 +11,7 @@ namespace NLayer.Core
         public string Name { get; set; }
         public int Stock { get; set; }
         public decimal Price { get; set; }
-        public decimal CategoryId { get; set; }
+        public int CategoryId { get; set; }
 
         // 1 product can have only 1 category
         public Category Category { get; set; }
